Refactor TweetCollection class to have inner classes for Document
